requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
favicon==0.7.0
feedparser==6.0.8
idna==3.3
obs-websocket-py==0.5.3
Pillow==9.0.1
python-dotenv==0.19.2
requests==2.27.1
sgmllib3k==1.0.0
six==1.16.0
snakeviz==2.1.1
soupsieve==2.3.1
tornado==6.1
urllib3==1.26.9
websocket-client==1.3.1
